requirements.txt
crewai
crewai-tools
langchain
langchain-community
langchain-openai
langchain-groq
langchain-anthropic
langchain-ollama
streamlit
python-dotenv
pdfminer.six
sqlalchemy
psycopg2-binary
snowflake-connector-python
markdown
docling
azure.identity

openai 
python-dotenv